Balance textures waves (#227)
* Wave Spawn Radius/Asteroid health reduction

Lowered the health of asteroids as they will be the first thing a player is likely to shoot at. At 100 health the guns all felt quite anemic. Enemy waves now spawn out at the radius of 55 so they appear off-screen.

* Weapon Balance and descriptions

Provided Descriptions to poorly described weapons in the loadout menu

* Start wave button has a 5 second delay between clicks.

In future, will need to hide start wave button when it is not usable.

Co-authored-by: MistaMishMash <MistaMishMash@users.noreply.github.com>