Boss, Tutorials and ship selection bug fix (#360)
Hai! Liam M actually did code! (And totally wasn't told everything to do by another Liam).

Fixed the ship selection still selecting the wrong ship.
Added a boss fight to the game for when the doom timer caps.
Added a placeholder tutorial event that triggers on game start and gave the player the option to disable it in the options menu.