Move all the explosion stuff to a clean branch (#371)
* Move all the explosion stuff to a clean branch

* Add a ship-repair event, tutorial details and end-of-game details (#372)

Added a pair of events that both just repair your ship.  As we currently have no spawn weighting system, I've made two of the same event to make it more likely to show up than the hostile option.

Added a basic tutorial and end-of-game description to their respective events.

* Build interface QoL improvements (#370)

* Build hint text will now only show if you have modules in your inventory.

* floating text when obtaining modules is a different colour

* Module list is sorted alphabetically

* Updated module floating text colour to be lighter and easier to see

* Damage Resistance Modifier (#366)

* Changed value for Damage Resistance Modifier

Changed value for Damage Resistance Modifier

Still had it on -15 instead of -0.15

* Removed debug logs

Removed debug logs

* Title & Diaolgue(#365)

* Unselected raycast targets for Opening Screen.

Unselected raycast targets for Opening Screen.

This should fix the click through being unresponsive.

Press any key still needs work.

* Update Mainmenu.unity

Forgot to unselect raycast target on one of the effects.

* Added PressAnyKey.cs

Added PressAnyKey.cs to button in opening screen.
This will press the button to take you to the next screen if any key on the keyboard is pressed.

* Updated PressAnyKey.cs

Updated PressAnyKey.cs

Noticed you could activate the script without activating the animation.
Technically this shouldn't be possible but unity sucks.

* Set wait time to 0 for AnimationPlayer

This would be the main cause of the delay when trying to get through the main menu quickly as it would not play for the initial wait even when called.

* Update PressAnyKey.cs

* fixed dialogue sizing

Co-authored-by: Liam Harrison <liam.har98@gmail.com>

* Enemy spawn update (#364)

* Enemies will now spawn outside of camera view, and away from asteroids. Both values may be modified in WaveManager

* Commenting, cleaning up, making variable order more logical

* Fixed Attack Speed Module, modules can now update effects on Level Up (#361)

* fixed up attack speed system logic (so when it reverts it returns to original value)

* Updated charge weapon so its firing rate is proportional to RateOfFire and deltatime. Updated sunburst so its in-game firing rate is the same. Verified rateOfFire module works for all weapons.

* clean up

* Testing out OnLevelChanged, to stop weapon fire rate from having a net decrease if unequipped after a level up

* Level up issue has been resolved.

* Cleaned up, fire rate system is affecting fire rate of all weapons correctly, including when levelled up. It also reverts stats correctly when unequipped.

* Updated ship speed, shield, and tractor AoE systems to update their effects when they level up. All have been playtested.

* Updated tractor beam description to reflect current behaviour

* RareMetal > Module conversion (#359)

Currently notifies the player with the same pickup floating text as picking up other drops does.  This can be changed if needed!

Also changed a comment in NotificationLabel to direct to the correct script of PrintNotification.

* Move all the explosion stuff to a clean branch

* + 2 bugfixes

Particle system submesh bug, and the LootEntity nullref bug, hopefully squashed!

Co-authored-by: LiamCMoore <42259108+LiamCMoore@users.noreply.github.com>
Co-authored-by: MistaMishMash <MistaMishMash@users.noreply.github.com>
Co-authored-by: Anthony Tran <100575155@student.swin.edu.au>
Co-authored-by: Liam Harrison <liam.har98@gmail.com>